+++ src/routes/envman/portainer-sync.ts
@@ -1,4 +1,6 @@
 import { Elysia } from 'elysia'
+import { portainerLogsStreamRouter } from './portainer-logs-stream'
+import { portainerRestartRouter } from './portainer-restart'
 import { syncContainersRouter } from './portainer-sync-containers'
 import { syncCoreRouter } from './portainer-sync-core'
 import { syncLifecycleRouter } from './portainer-sync-lifecycle'
@@ -11,3 +13,5 @@ export const syncRouter = new Elysia()
   .use(syncLifecycleRouter)
   .use(syncOpsRouter)
   .use(syncContainersRouter)
+  .use(portainerRestartRouter)
+  .use(portainerLogsStreamRouter)